hrms: changes for yearofpassing

+++ rainmaker/egov-hrms/src/main/java/org/egov/hrms/model/DepartmentalTest.java
@@ -59,7 +59,7 @@ public class DepartmentalTest {
 	private String test;
 
 	@NotNull
-	private Integer yearOfPassing;
+	private Long yearOfPassing;
 
 	private String remarks;
